edit(wip): moved functions and removed relative imports
