+++ src/sm/app/anon/subcommands/manual.py
@@ -1,8 +1,8 @@
 import typer
 from typing import Annotated, Optional
-from ..funcs import anon_func
-from ...helper_funcs import return_flags, load_schema, load_file_path, close_folder
-from ...models import AnonymiserConfig
+from sm.app.anon.funcs import anon_func
+from sm.app.helper_funcs import  return_flags, load_schema, load_file_path, close_folder
+from sm.app.models import AnonymiserConfig
 
 import json
