+++ src/sm/app/synth/funcs.py
@@ -1,8 +1,7 @@
 import json
 import time
-from ..helper_funcs import send_batch_to_API
-
-from ...synthesiser.synthesiser import Synthesiser
+from sm.app.helper_funcs import send_batch_to_API
+from sm.synthesiser.synthesiser import Synthesiser
 
 
 def synth_func(
